feat: bump wasm-bindgen from 0.2.95 to 0.2.99
Although highly unlikely, there is a chance that when user clicks Go! button fast enough, ChordProgressionManager.loadConfig would be called when it is still executing ChordProgressionManager.calculateChordProgression. In the future when calculateChordProgression takes significantly long, this issue would surface much more often. Because the former takes a mut borrow self, while a self is already read-only-ly borrowed by calculateChordProgression, it would be dynamic violation to rust's borrow rules and get caught by wasm-bindgen in debug mode, throwing "recursive use of an object detected which would lead to unsafe aliasing in rust" error. In release mode, it would throw "memory access out of bounds" (essentially wasm-bindgen detects that it is trying to double-free a segment of wasm linear memory).

This can be easily fixed in JavaScript by adding a wrapper, but would be better to fix it directly in rust.